Finish the transition from the local to the global ClientController.
The ClientController now sends the user messages directly to the
SwarmManagers using the sendDirect function. The gate userCommand was
added to the SwarmManager, and is the entry point for the
ClientController messages. The previous version used direct method
calling. Also, update the configuration files to work with the new
architecture.