Fix the state machines to correctly close the TCP connection from both sides.
Before the correction, the connection was being closed with abort(),
which is not optimal. Now, the connection is closed when both the client
and the peer issue a close() command. Also, the client now identifies
when the connection was remotely closed because of a refusal during the
handshake sequence and acts accordingly.