Move the Rate Limiter inside the ContentManager.
The external RateLimiter model was having a great impact on
the simulation performance because of the great number of PieceMsgs that
were being queued.

The threads now request the piece to the ContentManager, which in turn
will warn the thread via an application message that the piece is ready
for sending.

This commit will probably break execution.